use floating point math in sensor fusion

+++ modules/src/sensfusion6.c
@@ -240,9 +240,9 @@ void sensfusion6GetEulerRPY(float* roll, float* pitch, float* yaw)
   if (gx>1) gx=1;
   if (gx<-1) gx=-1;
 
-  *yaw = atan2(2*(q0*q3 + q1*q2), q0*q0 + q1*q1 - q2*q2 - q3*q3) * 180 / M_PI;
-  *pitch = asin(gx) * 180 / M_PI; //Pitch seems to be inverted
-  *roll = atan2(gy, gz) * 180 / M_PI;
+  *yaw = atan2f(2*(q0*q3 + q1*q2), q0*q0 + q1*q1 - q2*q2 - q3*q3) * (float) (180 / M_PI);
+  *pitch = asinf(gx) * (float) (180 / M_PI); //Pitch seems to be inverted
+  *roll = atan2f(gy, gz) * (float) (180 / M_PI);
 }
 
 float sensfusion6GetAccZWithoutGravity(const float ax, const float ay, const float az)
@@ -255,7 +255,7 @@ float sensfusion6GetAccZWithoutGravity(const float ax, const float ay, const flo
 
   // return vertical acceleration without gravity
   // (A dot G) / |G| - 1G (|G| = 1) -> (A dot G) - 1G
-  return ((ax*gx + ay*gy + az*gz) - 1.0);
+  return ((ax*gx + ay*gy + az*gz) - 1.0f);
 }
 //---------------------------------------------------------------------------------------------------
 // Fast inverse square-root
